+++ src/plugins/map_generation/src/traits/key_mapper.rs
@@ -1,5 +1,5 @@
 use bevy::prelude::*;
 
 pub trait KeyMapper {
-	fn key_for(&self, translation: Vec3) -> Option<(usize, usize)>;
+	fn key_for(&self, translation: Vec3) -> Option<(u32, u32)>;
 }
